dental_agent/requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic[email]

# Database
sqlmodel>=0.0.14
sqlalchemy>=2.0.0

# Background Tasks (Celery + Redis)
celery>=5.3.0
redis>=5.0.0

# Authentication
PyJWT>=2.8.0
python-dotenv>=1.0.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# WebSocket Client (for Deepgram Voice Agent)
websockets>=12.0

# AI/Voice (Deepgram)
deepgram-sdk>=3.0.0

# OpenAI (for LLM in Voice Agent)
openai>=1.0.0

# Telephony
twilio>=8.0.0

# Data Processing
pandas>=2.0.0
python-multipart>=0.0.6

# Validation & Security
phonenumbers>=8.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
aiofiles>=23.0.0

# Local development
pyngrok>=7.0.0
